beta-13.8

+++ stm32f103/src/sd_card/sdcard.cpp
@@ -175,7 +175,7 @@ void Logger::logKV(const char* type, int count, ...) {
 
     for (int i = 0; i < count; i++) {
         const char* key = va_arg(args, const char*);
-        float value = va_arg(args, double); // float → double
+        double value = va_arg(args, double);
         const char* unit = va_arg(args, const char*);
 
         offset += sprintf(buffer + offset, "%s=%.2f%s ", key, value, unit);
@@ -205,7 +205,7 @@ void Logger::logMixed(const char* type, const char* message, int count, ...) {
 
     for (int i = 0; i < count; i++) {
         const char* key = va_arg(args, const char*);
-        float value = va_arg(args, double);
+        double value = va_arg(args, double);
         const char* unit = va_arg(args, const char*);
 
         offset += sprintf(buffer + offset, "%s=%.2f%s ", key, value, unit);
